feat(backend): nueva logica de fuente relaciones mdea y ods funcionando

+++ src/main/java/com/seleccion/backend/entities/variables/variable_revision_masiva_update_dto.java
@@ -9,7 +9,7 @@ import lombok.*;
 @Builder
 public class variable_revision_masiva_update_dto {
     private List<String> idsA;
-    private Short prioridad;
+    private Integer prioridad;
     private Boolean revisada;
     private Integer responsableRevision;
 }
